refactor: extract super ops internal snapshot
